+++ Library/src/main/java/com/programmers/library/service/LibraryService.java
@@ -7,14 +7,9 @@ import com.programmers.library.exception.ExceptionHandler;
 import com.programmers.library.repository.Repository;
 
 import java.util.List;
-import java.util.Timer;
-import java.util.TimerTask;
-
-import static com.programmers.library.domain.BookStatusType.RENTABLE;
 
 public class LibraryService {
     private final Repository repository;
-    private static final int ORGANIZING_TIME = 5 * 60 * 1000;
 
     public LibraryService(Repository repository) {
         this.repository = repository;
@@ -46,16 +41,4 @@ public class LibraryService {
         repository.findBookById(id).orElseThrow(() -> ExceptionHandler.err(ErrorCode.BOOK_NOT_FOUND));
         repository.deleteBook(id);
     }
-
-    public void completeOrganizing(Book book) {
-        Timer timer = new Timer();
-        timer.schedule(new TimerTask() {
-            @Override
-            public void run() {
-                // 5분 후에 실행되는 코드
-                repository.updateStatus(book, RENTABLE);
-            }
-        }, ORGANIZING_TIME); // 5분(밀리초 단위)
-    }
-
 }
